requirements.txt
numpy
pandas
matplotlib
requests
python-binance
Django
streamlit